Moved `Badge` outside of the code directory
